fix: Configuration initialization should be done before the springboot startup bean is loaded
