+++ streampark-console/streampark-console-service/src/main/java/org/apache/streampark/console/StreamParkConsoleBootstrap.java
@@ -17,14 +17,14 @@
 
 package org.apache.streampark.console;
 
+import org.apache.streampark.console.core.runner.EnvApplicationContextInitializer;
+
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.scheduling.annotation.EnableScheduling;
 
 /**
- *
- *
  * <pre>
  *
  *      _____ __                                             __
@@ -46,7 +46,9 @@ import org.springframework.scheduling.annotation.EnableScheduling;
 @EnableScheduling
 public class StreamParkConsoleBootstrap {
 
-  public static void main(String[] args) {
-    SpringApplication.run(StreamParkConsoleBootstrap.class, args);
-  }
+    public static void main(String[] args) {
+        new SpringApplicationBuilder(StreamParkConsoleBootstrap.class)
+            .initializers(new EnvApplicationContextInitializer())
+            .run(args);
+    }
 }
