+++ streampark-console/streampark-console-service/src/main/java/org/apache/streampark/console/core/runner/EnvInitializer.java
@@ -20,7 +20,6 @@ package org.apache.streampark.console.core.runner;
 import org.apache.streampark.common.conf.CommonConfig;
 import org.apache.streampark.common.conf.ConfigKeys;
 import org.apache.streampark.common.conf.InternalConfigHolder;
-import org.apache.streampark.common.conf.InternalOption;
 import org.apache.streampark.common.conf.Workspace;
 import org.apache.streampark.common.enums.StorageType;
 import org.apache.streampark.common.fs.FsOperator;
@@ -41,7 +40,6 @@ import org.springframework.boot.ApplicationArguments;
 import org.springframework.boot.ApplicationRunner;
 import org.springframework.context.ApplicationContext;
 import org.springframework.core.annotation.Order;
-import org.springframework.core.env.Environment;
 import org.springframework.stereotype.Component;
 
 import java.io.File;
@@ -50,7 +48,6 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.util.Arrays;
 import java.util.HashSet;
-import java.util.Optional;
 import java.util.Set;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -78,25 +75,8 @@ public class EnvInitializer implements ApplicationRunner {
   @SneakyThrows
   @Override
   public void run(ApplicationArguments args) throws Exception {
-    Optional<String> profile =
-        Arrays.stream(context.getEnvironment().getActiveProfiles()).findFirst();
-    if ("test".equals(profile.orElse(null))) {
-      return;
-    }
-
-    String appHome = WebUtils.getAppHome();
-    if (StringUtils.isBlank(appHome)) {
-      throw new ExceptionInInitializerError(
-          String.format(
-              "[StreamPark] System initialization check failed,"
-                  + " The system initialization check failed. If started local for development and debugging,"
-                  + " please ensure the -D%s parameter is clearly specified,"
-                  + " more detail: https://streampark.apache.org/docs/user-guide/deployment",
-              ConfigKeys.KEY_APP_HOME()));
-    }
-
     // init InternalConfig
-    initInternalConfig(context.getEnvironment());
+    initInternalConfig();
     // overwrite system variable HADOOP_USER_NAME
     String hadoopUserName = InternalConfigHolder.get(CommonConfig.STREAMPARK_HADOOP_USER_NAME());
     overrideSystemProp(ConfigKeys.KEY_HADOOP_USER_NAME(), hadoopUserName);
@@ -106,17 +86,7 @@ public class EnvInitializer implements ApplicationRunner {
     ZIOExt.unsafeRun(EmbeddedFileServer.launch());
   }
 
-  private void initInternalConfig(Environment springEnv) {
-    // override config from spring application.yaml
-    InternalConfigHolder.keys().stream()
-        .filter(springEnv::containsProperty)
-        .forEach(
-            key -> {
-              InternalOption config = InternalConfigHolder.getConfig(key);
-              Utils.requireNotNull(config);
-              InternalConfigHolder.set(config, springEnv.getProperty(key, config.classType()));
-            });
-
+  private void initInternalConfig() {
     settingService.getMavenConfig().updateConfig();
 
     InternalConfigHolder.log();
